+++ apps/server/internal/store/db.go
@@ -57,7 +57,6 @@ func Open(cfg config.DatabaseConfig) (*gorm.DB, error) {
 		&model.FileOverrideSet{},
 		&model.FileOverrideSetRevision{},
 		&model.ZoneAssignment{},
-		&model.ZoneDefaultEntry{},
 		&model.ServerDrain{},
 		&model.ServerOffline{},
 		&model.AuditLog{},